fix: board js

+++ static/js/board.js
@@ -556,18 +556,27 @@ function editQuestion(id) {
 }
 
 function deleteAnswer(id) {
-    fetchWithToken(apiEndpoint + 'board_answers/' + id, {
-        method: 'DELETE',
-    })
-    .then(response => {
-        if (!response.ok) {
-            throw new Error('Network response was not ok');
-        }
-        getQuestion(currentQuestionId);
-    })
-    .catch(error => {
-        console.error('Error:', error);
-    });
+    // 'confirm' 함수로 사용자에게 확인 메시지를 표시합니다.
+    var confirmation = confirm("정말로 이 댓글을 삭제하시겠습니까?");
+
+    // 'confirm' 함수의 결과를 확인하여 삭제 작업을 계속할지 결정합니다.
+    if (confirmation) {
+        fetchWithToken(apiEndpoint + 'board_answers/' + id, {
+            method: 'DELETE',
+        })
+        .then(response => {
+            if (!response.ok) {
+                throw new Error('Network response was not ok');
+            }
+            getQuestion(currentQuestionId);
+        })
+        .catch(error => {
+            console.error('Error:', error);
+        });
+    } else {
+        // '취소' 버튼이 클릭되면, 아무런 동작을 수행하지 않습니다.
+        return;
+    }
 }
 
 function editAnswer(id) {
